[DTM] Adjusting marketing_offline_manual_costs (#4808)
* adjusting aux_key

* casting bigint

+++ bietlejuice/jobs/composer/db/datalake/queries/gsheets/clean/marketing_offline_manual_costs.sql
@@ -4,7 +4,7 @@ SELECT
     NULLIF(email, '') AS email,
     NULLIF(team, '') AS team,
     NULLIF(entry_type, '') AS entry_type,
-    INT(NULLIF(invoice_number, '')) AS invoice_number,
+    BIGINT(NULLIF(invoice_number, '')) AS invoice_number,
     NULLIF(cnpj, '') AS cnpj,
     NULLIF(vendor_name, '') AS vendor_name,
     NULLIF(description, '') AS description,
